add make bootstrap to fix error
Error is

Warning (emacs): Org version mismatch. Org loading aborted. This warning usually appears when a built-in Org version is loaded prior to the more recent Org version.

Version mismatch is commonly encountered in the following situations:

1. Emacs is loaded using literate Org config and more recent Org
   version is loaded inside the file loaded by ‘org-babel-load-file’.
   ‘org-babel-load-file’ triggers the built-in Org version clashing
   the newer Org version attempt to be loaded later.

   It is recommended to move the Org loading code before the
   ‘org-babel-load-file’ call.

2. New Org version is loaded manually by setting ‘load-path’, but
some other package depending on Org is loaded before the ‘load-path’
is configured. This "other package" is triggering built-in Org
version, again causing the version mismatch.

It is recommended to set ‘load-path’ as early in the config as
possible.

3. New Org version is loaded using straight.el package manager and
other package depending on Org is loaded before straight triggers
loading of the newer Org version.

It is recommended to put

(straight-use-package 'org)

early in the config.  Ideally, right after the straight.el
bootstrap.  Moving ‘use-package’ :straight declaration may not be
sufficient if the corresponding ‘use-package’ statement is
deferring the loading. Disable showing Disable logging
Warning (initialization): An error occurred while loading
‘/home/alessandro/.emacs.d/init.el’:

error: Org version mismatch.  Make sure that correct ‘load-path’ is
set early in init.el

To ensure normal operation, you should investigate and remove the
cause of the error in your initialization file.  Start Emacs with
the ‘--debug-init’ option to view a complete error backtrace.